+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Event handler register/unregister.
 - Read misreporting a closed connection as a timeout.
 - IpV4Address and IpV6Address comparison operators not being const.
+- sprintf used instead of snprintf for address formatting.
 
 ## [1.2.2] - 2026-08-11
 ### Fixed